requirements.txt
Flask==2.2.2
flask-cors==3.1.1
firebase-admin==5.2.0
livekit==0.1.3
python-dotenv==1.0.0